Fix accuracy issues with time

+++ synth/src/clock.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, Copy)]
 pub struct Clock {
-    pub time: f64,
+    pub time: u64,
     pub sample_rate: u64,
 }
